pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "odinbot"
version = "0.1.0"
description = "A Discord bot that checks UUIDs in the ODIN Threatfeed API and can also analyze Discord messages"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.2",  # Discord bot functionality
    "httpx>=0.26.0",     # Async HTTP requests
    "loguru>=0.7.2",     # Logging
    "requests>=2.31.0",  # HTTP requests
    "python-dotenv>=1.0.0",  # Environment variables
    "any-agent>=0.1.0",  # Agent functionality
    "litellm>=1.30.7",   # LLM calls
    "fire>=0.5.0",       # CLI for agent
    "pydantic>=2.6.1",   # Data validation
    "click>=8.1.7",      # CLI framework
    "docker>=7.0.0"      # For MCPStdio tool
]

[project.scripts]
odinbot = "odinbot.cli:main"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 